requirements.txt
fastapi
uvicorn
streamlit
requests
langchain-cohere
python-dotenv
PyPDF2
python-docx
pandas
